+++ src/Sentry.Profiling/SampleProfilerSession.cs
@@ -56,19 +56,12 @@ internal class SampleProfilerSession : IDisposable
 
     public TraceLog TraceLog => EventSource.TraceLog;
 
-    // default is false, set 1 for true.
-    private static int _throwOnNextStartupForTests = 0;
+    private static InterlockedBoolean _throwOnNextStartupForTests = false;
 
     internal static bool ThrowOnNextStartupForTests
     {
-        get { return Interlocked.CompareExchange(ref _throwOnNextStartupForTests, 1, 1) == 1; }
-        set
-        {
-            if (value)
-                Interlocked.CompareExchange(ref _throwOnNextStartupForTests, 1, 0);
-            else
-                Interlocked.CompareExchange(ref _throwOnNextStartupForTests, 0, 1);
-        }
+        get { return _throwOnNextStartupForTests; }
+        set { _throwOnNextStartupForTests.Exchange(value); }
     }
 
     public static SampleProfilerSession StartNew(IDiagnosticLogger? logger = null)
@@ -77,7 +70,7 @@ internal class SampleProfilerSession : IDisposable
         {
             var client = new DiagnosticsClient(Environment.ProcessId);
 
-            if (Interlocked.CompareExchange(ref _throwOnNextStartupForTests, 0, 1) == 1)
+            if (_throwOnNextStartupForTests.CompareExchange(false, true) == true)
             {
                 throw new Exception("Test exception");
             }