+++ src/Sentry/TransactionTracer.cs
@@ -12,8 +12,9 @@ public class TransactionTracer : IBaseTracer, ITransactionTracer
     private readonly IHub _hub;
     private readonly SentryOptions? _options;
     private readonly Timer? _idleTimer;
-    private long _cancelIdleTimeout;
     private readonly SentryStopwatch _stopwatch = SentryStopwatch.StartNew();
+
+    private InterlockedBoolean _cancelIdleTimeout;
 
     private readonly Instrumenter _instrumenter = Instrumenter.Sentry;
 
@@ -247,7 +248,7 @@ public class TransactionTracer : IBaseTracer, ITransactionTracer
         // Set idle timer only if an idle timeout has been provided directly
         if (idleTimeout.HasValue)
         {
-            _cancelIdleTimeout = 1;  // Timer will be cancelled once, atomically setting this back to 0
+            _cancelIdleTimeout = true;  // Timer will be cancelled once, atomically setting this back to false
             _idleTimer = new Timer(state =>
             {
                 if (state is not TransactionTracer transactionTracer)
@@ -362,7 +363,7 @@ public class TransactionTracer : IBaseTracer, ITransactionTracer
     public void Finish()
     {
         _options?.LogDebug("Attempting to finish Transaction {0}.", SpanId);
-        if (Interlocked.Exchange(ref _cancelIdleTimeout, 0) == 1)
+        if (_cancelIdleTimeout.Exchange(false) == true)
         {
             _options?.LogDebug("Disposing of idle timer for Transaction {0}.", SpanId);
             _idleTimer?.Change(Timeout.InfiniteTimeSpan, Timeout.InfiniteTimeSpan);